+++ plugins/system/cwgears/cwgears.php
@@ -190,20 +190,22 @@ class plgSystemCwgears extends JPlugin {
         //Lets add some style for backend extension configurations.
         if ($app->isAdmin()) {
 
-            if ($option == 'com_categories' && ($ext == 'com_coalawebcomments' || $ext == 'com_coalawebmarket' || $ext == 'com_coalawebtraffic' || $ext == 'com_coalaweblingual' || $ext == 'com_coalawebmembers'|| $ext == 'com_coalawebvideo')) {
+            if ($option == 'com_categories' && ($ext == 'com_coalawebcomments' || $ext == 'com_coalawebmarket' || $ext == 'com_coalawebtraffic' || $ext == 'com_coalaweblingual' || $ext == 'com_coalawebmembers'|| $ext == 'com_coalawebvideo'|| $ext == 'com_coalawebimport')) {
 
                 $doc->addStyleSheet($baseUrl . "components/generic/css/com-coalaweb-base-v2.css");
                 $doc->addStyleSheet($baseUrl . "components/generic/css/com-coalaweb-base-j3.css");
 
             }
 
-            if (in_array($option, array('com_coalawebcontact', 'com_coalawebsociallinks', 'com_coalawebtraffic', 'com_coalawebmarket', 'com_coalawebpaypal', 'com_coalaweblingual', 'com_coalawebcomments', 'com_coalawebmembers', 'com_coalawebvideo'))) {
+            if (in_array($option, array('com_coalawebcontact', 'com_coalawebsociallinks', 'com_coalawebtraffic', 'com_coalawebmarket', 'com_coalawebpaypal', 'com_coalaweblingual', 'com_coalawebcomments', 'com_coalawebmembers', 'com_coalawebvideo', 'com_coalawebimport'))) {
 
                 $doc->addStyleSheet($baseUrl . "components/generic/css/com-coalaweb-base-v2.css");
                 $doc->addStyleSheet($baseUrl . "components/generic/css/com-coalaweb-base-j3.css");
 
             }
         }
+
+        $doc->addStyleSheet($baseUrl . "plugins/system/gears/fonts/style.css");
 
         //----------------------------------------------------------------------
         // Gzip Help